+++ FirmwareUSB/vusb.h
@@ -1,0 +1,34 @@
+/*
+ * vusb.h
+ *
+ *  Created on: 14.05.2010
+ *      Author: Mike Shatohin (brunql)
+ *     Project: AmbilightUSB
+ *
+ *  AmbilightUSB is very simple implementation of the backlight for a laptop
+ *
+ *  Copyright (c) 2010, 2011 Mike Shatohin, mikeshatohin [at] gmail.com
+ *
+ *  AmbilightUSB is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation, either version 2 of the License, or
+ *  (at your option) any later version.
+ *
+ *  AmbilightUSB is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License
+ *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ */
+ 
+#ifndef USB_H_
+#define USB_H_
+
+#include "usbdrv.h"
+
+extern void usbInit_FakeUsbDisconnect(void);
+
+#endif /* USB_H_ */